Move code into subfolder and add config directory
All code files are now located in the tomef folder. When starting
jupyter lab adding tomef will automatically switch into this folder.
The configuration file is now located in the config folder.
This will make it easier to support multiple configuration files.

Relative paths in the configuration file are now relative to the
root directory of the repository for all files in config.
If you specify a configuration file using an absolute path
the links are relative to the directory containing the config file.