Improve default tokenizer to no longer use runvars
So far the default tokenizer was built from functions that pass around a
runvars dictionary to communicate. Now, all functions are part of the
DefaultTokenizer class. This is possible though the use of the
add_method decorator in util.py. The latter can be used to add a
function to a class after its definition. Therefore, it is still
possible to define the functions step by step in the notebook and define
the complete function at the end.

Additionally, the configuration settings have been improved to be more
intuitive and a wrong setting will raise a ConfigException that includes
a list of possible settings.