Restructure embedding module
All functionality of an embedding model provided to other components is
now accessed through properties. This includes the model, vector_size
and embedding function of the base EmbeddingModel as well as the vocab of
the WordEmbeddingModel. This way nothing needs to be loaded or
initialized explicitly.

The embedding function is used by both word and phrase embedding models
and returns a flattened array or a vector_size times instances matrix.
Each model has to provide _load functions for all those properties
specified above.

The fiter of the word2vec module is now specific to that class and no
longer used by all WordEmbeddingModels. Its functionality has been moved
into a special filter class.