Add faster word2vec token filter
The previous filter used get_close_matches to find the best fitting
token. Since this was rather slow, the new filter only uses the first
version of a phrase that it finds. All subsequent versions are dropped
from the lookup dictionaries.

The old filter can still be used by setting the fast_filter option in
the embedding config to false.