Move files to new locations and split code
The modules config, data and util are now in base. The iterating
functionality from util has been moved to a separate module
(iterator). The modules base and widgetbase have been restructured into
display, printer, progress, selectors and tables in widgets.

The functionality of nbprint is similar to before, but the printing
using markdown headings (#) has been removed. Therefore, the mini option
in nbbox no longer exists.

The NotebookBox can still be accessed as widgets.nbbox and is to be used
as with nbbox(): The advantage of this approach is that it will
automatically remove the progress iterator before finishing the section.
This removes the sideeffect of a widget that can not be loaded when
saving and reopening the file.

The iterate function has been split up into multiple functions to improve
readability and obsolete parts have been removed.